tests/wpt/tests/html/browsers/windows/browsing-context-names/choose-_self-001.html

<!DOCTYPE html>
<meta charset="utf-8">
<title>HTML Test: Choose browsing context - the given name is '_self'</title>
<link rel="author" title="Intel" href="http://www.intel.com/">
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<div id="log"></div>
<iframe src="resources/choose-_self-001-iframe.html" style="display:none"></iframe>
<script>
async_test(t => {
  window.addEventListener('message', t.step_func_done(e => {
    assert_equals(e.data.name, 'myownself');
  }), false);
}, 'The current browsing context must be chosen if the given name is "_self"');
</script>
